+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Possibility to edit the due date without resetting it first #52
+- Sort filters alphabetically #50
 
 ## [0.10.1] - 2024-09-22